gestion_reclamations/requirements.txt
Django==6.0.3
djangorestframework==3.17.0
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
psycopg2-binary==2.9.11
PyJWT==2.12.1
Pillow==12.1.1
